+++ howso/direct/core.py
@@ -2395,137 +2395,6 @@ class HowsoCore:
             "weight_feature": weight_feature,
         })
 
-    def pairwise_distances(
-        self,
-        trainee_id: str,
-        features: Optional[Iterable[str]] = None,
-        *,
-        action_feature: Optional[str] = None,
-        from_case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        from_values: Optional[List[List[Any]]] = None,
-        to_case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        to_values: Optional[List[List[Any]]] = None,
-        use_case_weights: bool = False,
-        weight_feature: Optional[str] = None
-    ) -> List:
-        """
-        Compute pairwise distances between specified cases.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        features : iterable of str, optional
-            List of feature names to use when computing pairwise distances.
-            If unspecified uses all features.
-        action_feature : str, optional
-            The action feature. If specified, uses targeted hyperparameters
-            used to predict this `action_feature`, otherwise uses targetless
-            hyperparameters.
-        from_case_indices : Iterable of Sequence[Union[str, int]], optional
-            An iterable of sequences, of session id and index, where index
-            is the original 0-based index of the case as it was trained into
-            the session. If specified must be either length of 1 or match
-            length of `to_values` or `to_case_indices`.
-        from_values : list of list of object, optional
-            A 2d-list of case values. If specified must be either length of
-            1 or match length of `to_values` or `to_case_indices`.
-        to_case_indices : Iterable of Sequence[Union[str, int]], optional
-            An Iterable of Sequences, of session id and index, where index
-            is the original 0-based index of the case as it was trained into
-            the session. If specified must be either length of 1 or match
-            length of `from_values` or `from_case_indices`.
-        to_values : list of list of object, optional
-            A 2d-list of case values. If specified must be either length of
-            1 or match length of `from_values` or `from_case_indices`.
-        use_case_weights : bool, default False
-            If set to True, will scale influence weights by each case's
-            `weight_feature` weight.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-
-        Returns
-        -------
-        list
-            A list of computed pairwise distances between each corresponding
-            pair of cases in `from_case_indices` and `to_case_indices`.
-        """
-        return self._execute(trainee_id, "pairwise_distances", {
-            "features": features,
-            "action_feature": action_feature,
-            "from_case_indices": from_case_indices,
-            "from_values": from_values,
-            "to_case_indices": to_case_indices,
-            "to_values": to_values,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-        })
-
-    def distances(
-        self,
-        trainee_id: str,
-        features: Optional[Iterable[str]] = None,
-        *,
-        action_feature: Optional[str] = None,
-        case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        feature_values: Optional[Iterable[Any]] = None,
-        use_case_weights: bool = False,
-        weight_feature: Optional[str] = None,
-        row_offset: int = 0,
-        row_count: Optional[int] = None,
-        column_offset: int = 0,
-        column_count: Optional[int] = None
-    ) -> Dict:
-        """
-        Compute distances matrix for specified cases.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        features : iterable of str, optional
-            List of feature names to use when computing distances. If
-            unspecified uses all features.
-        action_feature : str, optional
-            The action feature. If specified, uses targeted hyperparameters
-            used to predict this `action_feature`, otherwise uses targetless
-            hyperparameters.
-        case_indices : Iterable of Sequence[Union[str, int]], optional
-            An Iterable of Sequences, of session id and index, where index is
-            the original 0-based index of the case as it was trained into the
-            session. If specified, returns distances for all of these
-            cases. Ignored if `feature_values` is provided. If neither
-            `feature_values` nor `case_indices` is specified, uses full dataset.
-        feature_values : list of object, optional
-            If specified, returns distances of the local model relative to
-            these values, ignores `case_indices` parameter.
-        use_case_weights : bool, default False
-            If set to True, will scale influence weights by each case's
-            `weight_feature` weight.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-
-        Returns
-        -------
-        dict
-            A dictionary of keys, 'distances', 'row_case_indices' and
-            'column_case_indices'.
-        """
-        return self._execute(trainee_id, "distances", {
-            "features": features,
-            "action_feature": action_feature,
-            "case_indices": case_indices,
-            "feature_values": feature_values,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-            "row_offset": row_offset,
-            "row_count": row_count,
-            "column_offset": column_offset,
-            "column_count": column_count,
-        })
-
     def reset_parameter_defaults(self, trainee_id: str) -> None:
         """
         Reset Trainee hyperparameters and thresholds.
